Added support for Compute Module 6 parallel chains
 * Uses all available GPIO from 2 to 45.
 * Uses uint64 GPIO operations for the extended bits
   (regular Raspberry Pi only provides GPIO below 31).

+++ lib/framebuffer-internal.h
@@ -29,12 +29,12 @@ class RowAddressSetter;
 // An opaque type used within the framebuffer that can be used
 // to copy between PixelMappers.
 struct PixelDesignator {
-  PixelDesignator() : gpio_word(-1), r_bit(0), g_bit(0), b_bit(0), mask(~0){}
-  int gpio_word;
-  uint32_t r_bit;
-  uint32_t g_bit;
-  uint32_t b_bit;
-  uint32_t mask;
+  PixelDesignator() : gpio_word(-1), r_bit(0), g_bit(0), b_bit(0), mask(~0u){}
+  long gpio_word;
+  uint64_t r_bit;
+  uint64_t g_bit;
+  uint64_t b_bit;
+  uint64_t mask;
 };
 
 class PixelDesignatorMap {
